Port BunyipsLib changes

+++ TeamCode/src/au/edu/sa/mbhs/studentrobotics/bunyipslib/roadrunner/messages/PinpointLocalizerParamsMessage.java
@@ -24,9 +24,9 @@ public final class PinpointLocalizerParamsMessage {
      */
     public int deviceVersion;
     /**
-     * Scaling yaw factor used by the Pinpoint;
+     * Scaling yaw factor used by the Pinpoint.
      */
-    public float yawScalar;
+    public double yawScalar;
 
     @SuppressWarnings("MissingJavadoc")
     public PinpointLocalizerParamsMessage(@NonNull PinpointLocalizer.Params params, int deviceId, int deviceVersion, float yawScalar) {
